Split ProjectConfig up into three modules
Moved the config types into a separate module, and similarly with all
the legacy types and config parsing stuff.

Makes things a bit shorter and easier to grok.
